Move PreviousMarks out of InteractionLayer
