+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/InteractionLayer/InteractionLayerContainer.spec.js
@@ -2,7 +2,6 @@ import { shallow } from 'enzyme'
 import React from 'react'
 import InteractionLayerContainer from './InteractionLayerContainer'
 import InteractionLayer from './InteractionLayer'
-import PreviousMarks from './components/PreviousMarks'
 import SHOWN_MARKS from '@helpers/shownMarks'
 
 describe('Component > InteractionLayerContainer', function () {
@@ -39,17 +38,6 @@ describe('Component > InteractionLayerContainer', function () {
       />
     )
     expect(wrapper).to.be.ok()
-  })
-
-  it('should render PreviousMarks',  function () {
-    const wrapper = shallow(
-      <InteractionLayerContainer.wrappedComponent
-        activeInteractionTask={drawingTask}
-        height={height}
-        width={width}
-      />
-    )
-    expect(wrapper.find(PreviousMarks)).to.have.lengthOf(1)
   })
 
   describe('with an active drawing task and drawing tool', function () {
