+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/PreviousMarks/PreviousMarks.spec.js
@@ -1,7 +1,7 @@
 import { shallow } from 'enzyme'
 import React from 'react'
 import PreviousMarks from './PreviousMarks'
-import DrawingToolMarks from '../DrawingToolMarks'
+import DrawingToolMarks from '../InteractionLayer/components/DrawingToolMarks'
 import SHOWN_MARKS from '@helpers/shownMarks'
 
 describe('Component > PreviousMarks', function () {
